// src-31052018-currently-working/app/core/enums/sort-by-option.ts
/**
 * Enum for Sorting options
 */
export enum SortByOption {
    COMPLETION_PERCENTAGE = 'completionPercentage',
    EFFECTIVE_DATE = 'effectiveDate',
    CUSTOMER_NAME = 'customerName',
    PLAN_NAME = 'planName',
    PRODUCT = 'product',
    LAST_UPDATED = 'lastUpdatedTimestamp',
    CUSTOMER_NUMBER = 'customerNumber',
    LAST_MODIFIED = 'lastActivity',
    STATUS = 'status',
    FLAG_COUNT = 'flagsCount'
}